pyproject.toml
[tool.poetry]
name = "tvremote"
version = "0.1.0"
description = "A simple LG Smart TV Remote"
authors = ["Dimitris Strantsalis <dstrants@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.62.0"
uvicorn = "^0.13.1"
pywebostv = "^0.8.4"
PyYAML = "^5.3.1"
tinydb = "^4.3.0"

[tool.poetry.dev-dependencies]
pylama = "^7.7.1"
isort = "^5.6.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
